requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Audio visualizer dependencies
sounddevice>=0.4.6
numpy>=1.24.0

# Music file visualizer dependencies
soundfile>=0.12.1

# Video visualizer dependencies
opencv-python>=4.8.0

# Stock market visualizer dependencies
growwapi>=1.0.0
python-dotenv>=1.0.0
